Split large API into separate classes.
