+++ lib/mongo_browser/api.rb
@@ -1,125 +1,11 @@
 require "grape"
 require "grape-entity"
 
-module MongoBrowser
-  class Api < Grape::API
-    format :json
-    rescue_from :all, backtrace: true
+require "mongo_browser/entities"
+require "mongo_browser/api/documents"
+require "mongo_browser/api/collections"
+require "mongo_browser/api/databases"
+require "mongo_browser/api/mongo"
 
-    before do
-      method = env["REQUEST_METHOD"]
-      path = env["PATH_INFO"]
-
-      logger.info "[api] #{method} #{path}"
-    end
-
-    helpers do
-      def server
-        @server ||= MongoBrowser::Models::Server.current
-      end
-
-      def logger
-        Grape::API.logger
-      end
-    end
-
-    resource :databases do
-      desc "Get a list of all databases for the current server"
-      get do
-        present server.databases, with: Api::Entities::Database
-      end
-
-      params do
-        requires :db_name, type: String, desc: "Database name"
-      end
-      segment "/:db_name" do
-        desc "Deletes a database with the given name"
-        delete do
-          database = server.database(params[:db_name])
-          database.drop!
-          { success: true }
-        end
-
-        desc "Get stats for the given database"
-        get "/stats" do
-          database = server.database(params[:db_name])
-          database.stats
-        end
-
-        resources :collections do
-          desc "Get a list of all collections for the given database"
-          get do
-            database = server.database(params[:db_name])
-            collections = database.collections
-            present collections, with: Api::Entities::Collection
-          end
-
-          params do
-            requires :collection_name, type: String, desc: "Collection name"
-          end
-          segment "/:collection_name" do
-            desc "Get stats for a collection with the given name"
-            get "/stats" do
-              collection = server.database(params[:db_name]).collection(params[:collection_name])
-              collection.stats
-            end
-
-            desc "Drop a collection with the given name"
-            delete do
-              collection = server.database(params[:db_name]).collection(params[:collection_name])
-              collection.drop!
-              { success: true }
-            end
-
-            resources :documents do
-              desc "Get a list of paginated documents"
-              params do
-                optional :page, type: Integer, desc: "Page number"
-              end
-              get do
-                collection = server.database(params[:db_name]).collection(params[:collection_name])
-                documents = collection.documents_with_pagination(params[:page])
-                present documents, with: Api::Entities::PagedDocuments
-              end
-
-              params do
-                requires :id, type: String, desc: "Document id"
-              end
-              segment "/:id" do
-                desc "Get a document"
-                get do
-                  collection = server.database(params[:db_name]).collection(params[:collection_name])
-                  document = collection.find(params[:id])
-
-                  error!("Document not found", 404) if document.nil?
-                  present document, with: Api::Entities::Document
-                end
-
-                desc "Deletes a document with the given id"
-                delete do
-                  collection = server.database(params[:db_name]).collection(params[:collection_name])
-                  document = collection.find(params[:id])
-                  collection.remove!(document)
-                  { success: true }
-                end
-              end
-            end
-          end
-        end
-      end
-    end
-
-    desc "Returns info about the server"
-    get "/server_info" do
-      server.info
-    end
-
-    desc "Returns application version"
-    get "/version" do
-      {
-          version: MongoBrowser::VERSION,
-          environment: ENV["RACK_ENV"]
-      }
-    end
-  end
+module MongoBrowser::Api
 end
