+++ spec/lib/entities_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-describe MongoBrowser::Api::Entities::Database do
+describe MongoBrowser::Entities::Database do
   let(:database) { double }
   subject(:database_entity) { described_class.new(database) }
 
@@ -9,7 +9,7 @@ describe MongoBrowser::Api::Entities::Database do
   it { should expose :count }
 end
 
-describe MongoBrowser::Api::Entities::Collection do
+describe MongoBrowser::Entities::Collection do
   let(:collection) { double }
   subject(:database_entity) { described_class.new(collection) }
 
@@ -18,7 +18,7 @@ describe MongoBrowser::Api::Entities::Collection do
   it { should expose :size }
 end
 
-describe MongoBrowser::Api::Entities::Document do
+describe MongoBrowser::Entities::Document do
   let(:document) { double }
   subject(:database_entity) { described_class.new(document) }
 
@@ -26,7 +26,7 @@ describe MongoBrowser::Api::Entities::Document do
   it { should expose :data }
 end
 
-describe MongoBrowser::Api::Entities::PagedDocuments do
+describe MongoBrowser::Entities::PagedDocuments do
   let(:pager) { double }
   subject(:database_entity) { described_class.new(pager) }
